Improvements to game flow
Add Bezier Paths for the Moths
Add Debug Controls for changing the Paths
Add the Cat's Tail Animation
Add the Moth Idles to the Frontend
Only allow the pictures to be clicked on after the moths have flown away
Bring the Moths back after you click on each one
Allow the level to be restarted/reloaded with "R"
Allow the game to be debug "Paused"
Improved working with Flags. Can now set/unset multiple at the same
time.
Still need to figure out how to check mutliple at once.

Create a test level to help with the moths and curves.